Moved estimate_likelihood
